requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
openai
pydantic
python-dotenv